test: move several session specs to the main runner (#18433)
